Add storage-setup file and configuration
