+++ jestconfig.json
@@ -4,5 +4,6 @@
 },
 "testRegex": "(/__tests__/.*|(\\.|/)(test|spec))\\.(jsx?|tsx?)$",
 "moduleFileExtensions": ["ts", "tsx", "js", "jsx", "json", "node"],
-"testEnvironment":"jsdom"
+"testEnvironment":"jsdom",
+"setupFiles":["./test/storage-setup.ts"]
 }
