+++ test/storage/local/index.test.ts
@@ -5,10 +5,6 @@ import {
   toLocalStorage,
 } from '../../../src/storage/local';
 
-beforeEach(() => {
-  initLocalStorage();
-});
-
 describe('toLocalStorage', () => {
   it('should be defined', () => {
     expect(toLocalStorage).toBeDefined();
@@ -145,39 +141,3 @@ describe('clearLocalStorage', () => {
     expect(result2).toBeNull();
   });
 });
-
-function initLocalStorage() {
-  interface IStorage {
-    [key: string]: string;
-  }
-
-  const localStorageMock = (() => {
-    let store: IStorage = {};
-
-    return {
-      getItem(key: string) {
-        return store[key] || null;
-      },
-      setItem(key: string, value: any) {
-        store[key] = value;
-      },
-      removeItem(key: string) {
-        delete store[key];
-      },
-      clear() {
-        store = {};
-      },
-      get length() {
-        return Object.keys(store).length;
-      },
-      key(i: number) {
-        const keys = Object.keys(store);
-        return keys[i] || null;
-      },
-    };
-  })();
-
-  Object.defineProperty(window, 'localStorage', {
-    value: localStorageMock,
-  });
-}
